+++ src/content.ts
@@ -1,11 +1,10 @@
 import { requireGetCurrentFromURI, requireGetStore, ConfigOpts } from "./core/config";
 import {
-  isValidActivityContentNote,
-  isValidActivityContentProfile,
+  requireValidActivityContentNote,
+  requireValidActivityContentProfile,
   serialize,
   ActivityContentNote,
   ActivityContentProfile,
-  InvalidActivityContentError,
 } from "./core/activityContent";
 import {
   createBroadcast,
@@ -46,7 +45,7 @@ export const broadcast = async (
   contentObject: ActivityContentNote,
   opts?: ConfigOpts
 ): Promise<SignedBroadcastAnnouncement> => {
-  if (!isValidActivityContentNote(contentObject)) throw new InvalidActivityContentError();
+  requireValidActivityContentNote(contentObject);
   const content = serialize(contentObject);
 
   const currentFromURI = requireGetCurrentFromURI(opts);
@@ -91,7 +90,7 @@ export const reply = async (
 ): Promise<SignedReplyAnnouncement> => {
   if (!isDSNPAnnouncementURI(inReplyTo)) throw new InvalidAnnouncementUriError(inReplyTo);
 
-  if (!isValidActivityContentNote(contentObject)) throw new InvalidActivityContentError();
+  requireValidActivityContentNote(contentObject);
   const content = serialize(contentObject);
 
   const currentFromURI = requireGetCurrentFromURI(opts);
@@ -157,7 +156,7 @@ export const profile = async (
   contentObject: ActivityContentProfile,
   opts?: ConfigOpts
 ): Promise<SignedProfileAnnouncement> => {
-  if (!isValidActivityContentProfile(contentObject)) throw new InvalidActivityContentError();
+  requireValidActivityContentProfile(contentObject);
   const content = serialize(contentObject);
 
   const currentFromURI = requireGetCurrentFromURI(opts);